Add BibEntry method make_citekey()
Move make_entry_citekey in bibadd.py to method of BibEntry in bibfile.py,
refactor to reflect this change. Modify biblabel.py to use new method.

The new make_citekey function uses a slightly simpler label style dictionary.
It appears that the orginal dict format was an attempt to keep the style
definition similar to the styles in bibstuff.bibstyles, but make citation
labels for a BibTex database seems quite separate from producing in-text
citations and references as in the bibstyles subpackage and I did not see the
need for such generalization.

Also fix bibadd.py to remove pyflakes errors